finalize initial implementation
